fix: add withdray placeholder
